backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1

# Redis
redis==5.0.1
hiredis==2.3.2

# Vector store
chromadb==0.4.22

# LLM Providers
openai==1.10.0
anthropic==0.18.0
mistralai==0.1.3
together==0.2.8

# Embeddings
voyageai==0.2.0
cohere==4.47

# Document processing
pypdf2==3.0.1
python-docx==1.1.0
python-magic==0.4.27
pdfplumber==0.10.4

# Text processing
langchain==0.1.4
langchain-community==0.0.17
tiktoken==0.5.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
requests==2.31.0
aiohttp==3.9.1
tenacity==8.2.3

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Monitoring & Logging
loguru==0.7.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Data handling
pandas==2.1.4
numpy==1.26.3

# Local embeddings (BGE)
sentence-transformers==2.3.1
torch==2.1.2
